+++ sources/Decomposition/LQ.cs
@@ -16,8 +16,8 @@ namespace UMapx.Decomposition
     {
         #region Private data
         private QR qr;
-        private double[,] l;
-        private double[,] q;
+        private float[,] l;
+        private float[,] q;
         #endregion
 
         #region Initialize
@@ -25,7 +25,7 @@ namespace UMapx.Decomposition
         /// Initializes LQ decomposition.
         /// </summary>
         /// <param name="A">Square matrix</param>
-        public LQ(double[,] A)
+        public LQ(float[,] A)
         {
             qr = new QR(A.Transponate());
 
@@ -38,7 +38,7 @@ namespace UMapx.Decomposition
         /// <summary>
         /// Returns the lower triangular matrix L.
         /// </summary>
-        public double[,] L
+        public float[,] L
         {
             get
             {
@@ -48,7 +48,7 @@ namespace UMapx.Decomposition
         /// <summary>
         /// Returns the orthogonal matrix Q.
         /// </summary>
-        public double[,] Q
+        public float[,] Q
         {
             get
             {
